% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFDM.R
\name{AFDM}
\alias{AFDM}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
AFDM(
  data,
  InitialWt,
  FinalWt,
  Frac.InitialWt,
  Frac.FinalWt,
  Treatment,
  Day,
  Replicate
)
}
\arguments{
\item{data}{}

\item{InitialWt}{}

\item{FinalWt}{}

\item{Frac.InitialWt}{}

\item{Frac.FinalWt}{}

\item{Treatment}{}

\item{Day}{}

\item{Replicate}{}
}
\value{
Remaining
}
\description{
Calculate AFDM
}
\examples{
AFDM(data,InitialWt,FinalWt,Frac.InitialWt,Frac.FinalWt,Treatment,Day,Replicate)

}
