% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{manipulation}
\alias{manipulation}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
manipulation(data, InitialWt, FinalWt, Treatment, ...)
}
\arguments{
\item{data}{}

\item{InitialWt}{}

\item{FinalWt}{}

\item{Treatment}{}
}
\value{
meanControl
}
\description{
Calculate mass loss by transport to the field and manipulation
}
\examples{
manipulation(data,InitialWt,FinalWt,Treatment)

}
