% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotC.AFDM.R
\name{plot.C}
\alias{plot.C}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
\method{plot}{C}(data)
}
\arguments{
\item{data}{}

\item{Treatment}{}

\item{Replicate}{}

\item{Day}{}

\item{AFDMrem}{}
}
\description{
Made a variety of Plots to explore the data
}
\examples{
plot.C(data)

}
