% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD.AFDM.R
\name{plot.D}
\alias{plot.D}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
\method{plot}{D}(data)
}
\arguments{
\item{data}{}

\item{Treatment}{}

\item{Replicate}{}

\item{Day}{}

\item{AFDMrem}{}
}
\description{
Made a variety of Plots to explore the data
}
\examples{
plot.D(data)

}
