% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotE.AFDM.R
\name{plot.E}
\alias{plot.E}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
\method{plot}{E}(data)
}
\arguments{
\item{data}{}

\item{Treatment}{}

\item{Replicate}{}

\item{Day}{}

\item{AFDMrem}{}
}
\description{
Made a variety of Plots to explore the data
}
\examples{
plot.E(data)

}
