% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.k.R
\name{rsquared.k}
\alias{rsquared.k}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
rsquared.k(data, Treatment, Replicate, Day, Ln.AFDMrem)
}
\arguments{
\item{data}{}

\item{Treatment}{}

\item{Replicate}{}

\item{Day}{}

\item{Ln.AFDMrem}{}
}
\description{
Calculate r-squared
}
\examples{
rsquared.k(data,Treatment,Replicate,Day,Ln.AFDMrem)

}
