% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.k.R
\name{slope.k}
\alias{slope.k}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
slope.k(data, Treatment, Replicate, Day, Ln.AFDMrem)
}
\arguments{
\item{data}{}

\item{Treatment}{}

\item{Replicate}{}

\item{Day}{}

\item{Ln.AFDMrem}{}
}
\description{
Calculate slope
}
\examples{
slope.K <- function(data,Treatment,Replicate,Day,Ln.AFDMrem)

}
