% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA.AFDM.R, R/plotC.AFDM.R, R/plotD.AFDM.R,
%   R/plotE.AFDM.R
\name{plot.A}
\alias{plot.A}
\title{Leaf Litter Decomposition Analysis for Stream Ecosystems}
\usage{
\method{plot}{A}(data)

\method{plot}{A}(data)

\method{plot}{A}(data)

\method{plot}{A}(data)
}
\arguments{
\item{data}{}

\item{Treatment}{}

\item{Replicate}{}

\item{Day}{}

\item{AFDMrem}{}
}
\description{
Made a variety of Plots to explore the data

Made a variety of Plots to explore the data

Made a variety of Plots to explore the data

Made a variety of Plots to explore the data
}
\examples{
plot.A(data)

plot.A(data)

plot.A(data)

plot.A(data)

}
